# backend/public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion des Quotes-Parts - Authentification",
        "description": "Documentation des routes d’authentification (login, logout, me).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/banques": {
            "get": {
                "tags": [
                    "Banques"
                ],
                "summary": "Lister toutes les banques",
                "description": "Retourne la liste complète des banques.",
                "operationId": "733d354b91d57c9afe5fa5952ec1bf43",
                "responses": {
                    "200": {
                        "description": "Liste des banques récupérée avec succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Banques"
                ],
                "summary": "Créer une nouvelle banque",
                "description": "Ajoute une nouvelle banque.",
                "operationId": "f65c32a37f0e7d378d570b6578dc4c52",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "BNQ_CODE",
                                    "BNQ_LIBELLE"
                                ],
                                "properties": {
                                    "BNQ_CODE": {
                                        "type": "string"
                                    },
                                    "BNQ_LIBELLE": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Banque créée avec succès"
                    },
                    "409": {
                        "description": "Une banque avec ce nom existe déjà"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/banques/{code}": {
            "get": {
                "tags": [
                    "Banques"
                ],
                "summary": "Afficher une banque",
                "description": "Retourne les détails d’une banque à partir de son code.",
                "operationId": "28793dbf6702a26838c2a0244faa17cb",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "banque trouvée"
                    },
                    "404": {
                        "description": "banque non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Banques"
                ],
                "summary": "Mettre à jour une banque",
                "description": "Modifie les informations d’une banque existante.",
                "operationId": "f1e9d153bcaa4e412a0a887b11dfde86",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "BNQ_CODE": {
                                        "type": "string"
                                    },
                                    "BNQ_LIBELLE": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Banque mise à jour avec succès"
                    },
                    "404": {
                        "description": "Banque non trouvée"
                    },
                    "409": {
                        "description": "Une banque avec ce nom existe déjà"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Banques"
                ],
                "summary": "Supprimer une banque",
                "description": "Supprime une banque existante par son code.",
                "operationId": "04e5527c4ec416c0cb6550fb97a47e77",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Banque supprimée avec succès"
                    },
                    "404": {
                        "description": "Banque non trouvée"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}